+++ src/main/java/com/tiketeer/Tiketeer/domain/member/controller/MemberController.java
@@ -1,5 +1,8 @@
 package com.tiketeer.Tiketeer.domain.member.controller;
 
+import com.tiketeer.Tiketeer.domain.member.usecase.GetMemberUseCase;
+import com.tiketeer.Tiketeer.domain.member.usecase.MemberChargePointUseCase;
+import com.tiketeer.Tiketeer.domain.member.usecase.MemberRegisterUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,9 +18,6 @@ import com.tiketeer.Tiketeer.domain.member.controller.dto.ChargePointResponseDto
 import com.tiketeer.Tiketeer.domain.member.controller.dto.GetMemberResponseDto;
 import com.tiketeer.Tiketeer.domain.member.controller.dto.MemberRegisterRequestDto;
 import com.tiketeer.Tiketeer.domain.member.controller.dto.MemberRegisterResponseDto;
-import com.tiketeer.Tiketeer.domain.member.service.MemberPointService;
-import com.tiketeer.Tiketeer.domain.member.service.MemberRegisterService;
-import com.tiketeer.Tiketeer.domain.member.service.MemberService;
 import com.tiketeer.Tiketeer.domain.member.service.dto.GetMemberCommandDto;
 import com.tiketeer.Tiketeer.domain.member.service.dto.MemberRegisterCommandDto;
 import com.tiketeer.Tiketeer.response.ApiResponse;
@@ -27,23 +27,23 @@ import jakarta.validation.Valid;
 @RestController
 @RequestMapping("/members")
 public class MemberController {
-	private final MemberRegisterService memberRegisterService;
-	private final MemberPointService memberPointService;
-	private final MemberService memberService;
+	private final MemberRegisterUseCase memberRegisterUseCase;
+	private final MemberChargePointUseCase memberChargePointUseCase;
+
+	private final GetMemberUseCase getMemberUseCase;
 
 	@Autowired
-	public MemberController(MemberRegisterService memberRegisterService, MemberPointService memberPointService,
-		MemberService memberService) {
-		this.memberRegisterService = memberRegisterService;
-		this.memberPointService = memberPointService;
-		this.memberService = memberService;
+	public MemberController(MemberRegisterUseCase memberRegisterUseCase, MemberChargePointUseCase memberChargePointUseCase, GetMemberUseCase getMemberUseCase) {
+        this.memberRegisterUseCase = memberRegisterUseCase;
+        this.memberChargePointUseCase = memberChargePointUseCase;
+        this.getMemberUseCase = getMemberUseCase;
 	}
 
 	@PostMapping("/register")
 	public ApiResponse<MemberRegisterResponseDto> registerMember(
 		final @Valid @RequestBody MemberRegisterRequestDto registerMemberDto) {
 		return ApiResponse.wrap(
-			MemberRegisterResponseDto.toDto(memberRegisterService.register(MemberRegisterCommandDto.builder()
+			MemberRegisterResponseDto.toDto(memberRegisterUseCase.register(MemberRegisterCommandDto.builder()
 				.email(registerMemberDto.getEmail())
 				.isSeller(registerMemberDto.getIsSeller())
 				.build()
@@ -56,7 +56,7 @@ public class MemberController {
 		@Valid @RequestBody ChargePointRequestDto request) {
 		// TODO: JWT 구현이 완료되면 SecurityContext를 통해 가져오는 것으로 대체
 		var email = "mock@mock.com";
-		var totalPoint = memberPointService.chargePoint(request.convertToCommandDto(memberId, email)).getTotalPoint();
+		var totalPoint = memberChargePointUseCase.chargePoint(request.convertToCommandDto(memberId, email)).getTotalPoint();
 		var result = ChargePointResponseDto.builder().totalPoint(totalPoint).build();
 		return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.wrap(result));
 	}
@@ -64,7 +64,7 @@ public class MemberController {
 	@GetMapping("/")
 	public ResponseEntity<ApiResponse<GetMemberResponseDto>> getMember() {
 		var email = "mock@mock.com";
-		var result = memberService.getMember(GetMemberCommandDto.builder().memberEmail(email).build());
+		var result = getMemberUseCase.get(GetMemberCommandDto.builder().memberEmail(email).build());
 		var responseBody = ApiResponse.wrap(GetMemberResponseDto.convertFromDto(result));
 		return ResponseEntity.status(HttpStatus.OK).body(responseBody);
 	}
