--- src/test/java/com/tiketeer/Tiketeer/domain/member/service/MemberServiceTest.java
@@ -1,194 +1,0 @@
-package com.tiketeer.Tiketeer.domain.member.service;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.time.LocalDateTime;
-import java.util.Date;
-import java.util.UUID;
-
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.tiketeer.Tiketeer.auth.jwt.JwtPayload;
-import com.tiketeer.Tiketeer.auth.jwt.JwtService;
-import com.tiketeer.Tiketeer.domain.member.Otp;
-import com.tiketeer.Tiketeer.domain.member.exception.InvalidOtpException;
-import com.tiketeer.Tiketeer.domain.member.exception.InvalidTokenException;
-import com.tiketeer.Tiketeer.domain.member.exception.MemberNotFoundException;
-import com.tiketeer.Tiketeer.domain.member.repository.MemberRepository;
-import com.tiketeer.Tiketeer.domain.member.repository.OtpRepository;
-import com.tiketeer.Tiketeer.domain.member.service.dto.GetMemberCommandDto;
-import com.tiketeer.Tiketeer.domain.member.service.dto.InitMemberPasswordWithOtpCommandDto;
-import com.tiketeer.Tiketeer.domain.member.service.dto.RefreshAccessTokenCommandDto;
-import com.tiketeer.Tiketeer.domain.member.service.dto.RefreshAccessTokenResultDto;
-import com.tiketeer.Tiketeer.domain.role.constant.RoleEnum;
-import com.tiketeer.Tiketeer.domain.role.repository.RoleRepository;
-import com.tiketeer.Tiketeer.testhelper.TestHelper;
-
-@Import({TestHelper.class})
-@SpringBootTest
-public class MemberServiceTest {
-	@Autowired
-	private TestHelper testHelper;
-
-	@Autowired
-	private MemberService memberService;
-
-	@Autowired
-	private RoleRepository roleRepository;
-
-	@Autowired
-	private MemberRepository memberRepository;
-
-	@Autowired
-	private OtpRepository otpRepository;
-
-	@Autowired
-	private PasswordEncoder passwordEncoder;
-
-	@Autowired
-	private JwtService jwtService;
-
-	@BeforeEach
-	void initTable() {
-		testHelper.initDB();
-	}
-
-	@AfterEach
-	void cleanTable() {
-		testHelper.cleanDB();
-	}
-
-	@Test
-	@DisplayName("유효하지 않은 OTP > MemberService.initPasswordWithOtp 호출 > 실패")
-	@Transactional
-	void initPasswordWithOtpFailBecauseInvalidOtp() {
-		// given
-		var mockEmail = "test@test.com";
-		var member = testHelper.createMember(mockEmail);
-
-		otpRepository.save(
-			Otp.builder()
-				.member(member)
-				.expiredAt(LocalDateTime.of(9999, 12, 31, 0, 0))
-				.build());
-
-		var invalidOtp = UUID.randomUUID();
-
-		Assertions.assertThatThrownBy(() -> {
-			// when
-			memberService.initPasswordWithOtp(InitMemberPasswordWithOtpCommandDto.builder().otp(invalidOtp).build());
-			// then
-		}).isInstanceOf(InvalidOtpException.class);
-	}
-
-	@Test
-	@DisplayName("유효한 OTP > MemberService.initPasswordWithOtp 호출 > 성공")
-	@Transactional
-	void initPasswordWithOtpSuccess() {
-		// given
-		var mockEmail = "test@test.com";
-		var member = testHelper.createMember(mockEmail);
-
-		var otp = otpRepository.save(
-			Otp.builder()
-				.member(member)
-				.expiredAt(LocalDateTime.of(9999, 12, 31, 0, 0))
-				.build());
-
-		var mockPwd = "1qwertasdcxz";
-
-		// when
-		memberService.initPasswordWithOtp(
-			InitMemberPasswordWithOtpCommandDto.builder().otp(otp.getPassword()).password(mockPwd).build());
-
-		// then
-		var memberAfterEmailAuthOpt = memberRepository.findByEmail(mockEmail);
-		Assertions.assertThat(memberAfterEmailAuthOpt.isPresent()).isTrue();
-
-		var memberAfterEmailAuth = memberAfterEmailAuthOpt.get();
-		Assertions.assertThat(memberAfterEmailAuth.isEnabled()).isTrue();
-		Assertions.assertThat(passwordEncoder.matches(mockPwd, memberAfterEmailAuth.getPassword())).isTrue();
-
-		Assertions.assertThat(otpRepository.findById(otp.getPassword()).isPresent()).isFalse();
-	}
-
-	@Test
-	@DisplayName("정상 토큰 > 재발급 > 재발급 확인")
-	void refreshAccessToken() {
-		// given
-		String refreshToken = jwtService.createToken(
-			new JwtPayload("test@gmail.com", RoleEnum.BUYER, new Date(System.currentTimeMillis())));
-
-		// when
-		RefreshAccessTokenResultDto refreshAccessTokenResultDto = memberService.refreshAccessToken(
-			RefreshAccessTokenCommandDto.builder().refreshToken(refreshToken).build());
-
-		// then
-		String accessToken1 = refreshAccessTokenResultDto.getAccessToken();
-		JwtPayload jwtPayload = jwtService.verifyToken(accessToken1);
-
-		Assertions.assertThat(accessToken1).isNotNull();
-		Assertions.assertThat(jwtPayload.email()).isEqualTo("test@gmail.com");
-	}
-
-	@Test
-	@DisplayName("refreshToken 만료 > 재발급 > 예외발생 확인")
-	void refreshAccessTokenFailByExpiredRefreshToken() {
-		// given
-		String refreshToken = jwtService.createToken(
-			new JwtPayload("test@gmail.com", RoleEnum.BUYER,
-				new Date(System.currentTimeMillis() - 30L * 24 * 60 * 60 * 1000)));
-
-		// when
-		// then
-		assertThrows(InvalidTokenException.class, () -> {
-			memberService.refreshAccessToken(
-				RefreshAccessTokenCommandDto.builder().refreshToken(refreshToken).build());
-		});
-	}
-
-	@Test
-	@DisplayName("정상 조건 > 멤버 조회 > 성공")
-	@Transactional
-	void getMemberSuccess() {
-		// given
-		var mockEmail = "test@test.com";
-		var memberInDb = testHelper.createMember(mockEmail);
-
-		var command = GetMemberCommandDto.builder().memberEmail(mockEmail).build();
-
-		// when
-		var member = memberService.getMember(command);
-
-		// then
-		Assertions.assertThat(member.getEmail()).isEqualTo(memberInDb.getEmail());
-		Assertions.assertThat(member.getPoint()).isEqualTo(memberInDb.getPoint());
-		Assertions.assertThat(member.getProfileUrl()).isEqualTo(memberInDb.getProfileUrl());
-		Assertions.assertThat(member.getCreatedAt()).isEqualTo(memberInDb.getCreatedAt());
-	}
-
-	@Test
-	@DisplayName("멤버가 존재하지 않음 > 멤버 조회 > 실패")
-	@Transactional(readOnly = true)
-	void getMemberFailNotFoundMember() {
-		// given
-		var mockEmail = "test@test.com";
-
-		var command = GetMemberCommandDto.builder().memberEmail(mockEmail).build();
-
-		Assertions.assertThatThrownBy(() -> {
-			// when
-			memberService.getMember(command);
-			// then
-		}).isInstanceOf(MemberNotFoundException.class);
-	}
-}
